requirements.txt
astroid==2.1.0
atomicwrites==1.2.1
attrs==18.2.0
bcrypt==3.1.5
beautifulsoup4==4.6.3
boto3==1.9.200
botocore==1.12.253
cachetools==3.0.0
certifi==2018.11.29
cffi==1.14.5
chardet==3.0.4
coverage==4.5.2
dnspython==1.16.0
docutils==0.15.2
google-api-python-client==1.7.6
google-auth==1.6.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googlemaps==3.0.2
httplib2==0.19.0
idna==2.8
isort==4.3.4
jmespath==0.9.5
jsonschema==2.6.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
mock==4.0.2
mongodb-migrations==0.7.0
more-itertools==5.0.0
oauth2client==4.1.3
oauthlib==3.1.0
pluggy==0.8.1
py==1.10.0
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycparser==2.19
PyJWT==1.7.1
pylint==2.2.2
pymongo==3.6.0
pytest==4.0.2
pytest-cov==2.6.0
pytest-ordering==0.6
python-dateutil==2.7.5
requests==2.21.0
requests-oauthlib==1.3.0
rsa==4.7
s3transfer==0.2.1
six==1.12.0
sparkpost==1.3.6
titlecase==1.1.1
typed-ast==1.4.3
uritemplate==3.0.0
urllib3==1.24.2
wrapt==1.11.0
